Move resultIcon  for ResultCard component
