+++ client/src/components/1_molecules/index.ts
@@ -18,5 +18,4 @@ export * from './proposeAmend'
 export * from './followText'
 export * from './unFollowText'
 export * from './resultBar'
-export * from './resultIcon'
 export * from './collapse'
