--- client/src/components/1_molecules/resultIcon/index.ts
@@ -1,1 +1,0 @@
-export * from './resultIcon'
