+++ client/src/components/2_organisms/resultEventCard/footer.spec.tsx
@@ -1,19 +1,11 @@
 import * as React from 'react'
 import { render } from 'react-testing-library'
 import { amendMock } from '../../../../../interfaces'
-import { ResultIcon } from './resultIcon'
+import { Footer } from './footer'
 
 it('should render a Collapse without', () => {
   const { container, getByText } = render(
-    <ResultIcon
-      results={{
-        upVotesCount: 1,
-        downVotesCount: 2,
-        indVotesCount: 3,
-        totalPotentialVotesCount: 6
-      }}
-      isConfliced={false}
-    />
+    <Footer amend={amendMock} isConfliced={false} />
   )
   expect(container).toBeTruthy()
   expect(getByText('1')).toBeTruthy()
